refactor: 서버 리팩토링 (yoo20370,hyunS00)
