+++ src/interfaces/Position.interface.ts
@@ -1,0 +1,5 @@
+export default interface Position {
+  x: number;
+  y: number;
+  z: number;
+}
